+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Fix capturing stdout on legacy Windows https://github.com/Textualize/rich/pull/2055
+- Fix capturing stdout on legacy Windows https://github.com/Textualize/rich/pull/2066
 
 ## [12.0.0] - 2022-03-10